Lot of changes to support attention supervisio, unk vectors, fasttext
- FastText is supported via new modelvocab class now
- more verbose reporting
- more flags to make life easier

+++ pythia/core/tasks/datasets/base_dataset.py
@@ -82,7 +82,7 @@ class BaseDataset(Dataset):
         input_image_features = Variable(input_image_features)
 
         if input_contexts is not None:
-            input_contexts = Variable(input_contexts.type(torch.LongTensor))
+            input_contexts = Variable(input_contexts)
 
         if self.use_cuda:
             obs = obs.cuda()
